replace `useInertOthers` with `useInert`
